# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lunme API",
    "version": "1.0.0",
    "description": "Swagger gerado automaticamente pelas rotas do Next.js"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/internal/prontuario/analise-paciente": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Gera uma avaliação baseada no prompt enviado",
        "description": "Recebe um prompt, concatena com instruções internas e envia ao modelo de IA para gerar uma avaliação estruturada.",
        "tags": [
          "Interno - Avaliações"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "example": "Paciente demonstra sinais de melhora no tratamento."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avaliação gerada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Avaliação detalhada gerada pelo modelo."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro — campo prompt não enviado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Título e autor são obrigatórios."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao gerar resposta do modelo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao gerar resposta do modelo."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/internal/register_pacientes/pre-pacientes": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Lista pré-pacientes ainda não habilitados de um psicólogo",
        "description": "Retorna todos os registros de pré-pacientes vinculados ao psicólogo informado, filtrando apenas os não habilitados.",
        "tags": [
          "Interno - Pré-Pacientes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "psicologoId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do psicólogo responsável",
            "example": "clu1x0a9g0001z0t7i1s2abc"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pré-pacientes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "prp_123"
                      },
                      "nome": {
                        "type": "string",
                        "example": "Maria Oliveira"
                      },
                      "email": {
                        "type": "string",
                        "example": "maria@email.com"
                      },
                      "endereco": {
                        "type": "string",
                        "example": "Rua das Flores, 123"
                      },
                      "nascimento": {
                        "type": "string",
                        "example": "1995-08-12"
                      },
                      "idade": {
                        "type": "number",
                        "example": 29
                      },
                      "cpf": {
                        "type": "string",
                        "example": "12345678900"
                      },
                      "telefone": {
                        "type": "string",
                        "example": "27999998888"
                      },
                      "emergencia": {
                        "type": "string",
                        "example": "Mãe - 27988887777"
                      },
                      "generoOrientacao": {
                        "type": "string",
                        "example": "Feminino"
                      },
                      "estadoCivil": {
                        "type": "string",
                        "example": "Solteira"
                      },
                      "origemConhecimento": {
                        "type": "string",
                        "example": "Instagram"
                      },
                      "preocupacao": {
                        "type": "string",
                        "example": "Ansiedade constante"
                      },
                      "motivoAtendimento": {
                        "type": "string",
                        "example": "Crises de ansiedade"
                      },
                      "experienciaAnterior": {
                        "type": "string",
                        "example": "Já fez terapia 1 ano"
                      },
                      "saudeFisica": {
                        "type": "string",
                        "example": "Boa"
                      },
                      "detalhesSaudeFisica": {
                        "type": "string",
                        "example": "Sem condições graves"
                      },
                      "medicamentos": {
                        "type": "string",
                        "example": "Nenhum"
                      },
                      "diagnosticoMental": {
                        "type": "string",
                        "example": "Não possui"
                      },
                      "historicoFamiliar": {
                        "type": "string",
                        "example": "Histórico de ansiedade na família"
                      },
                      "rotina": {
                        "type": "string",
                        "example": "Trabalho intenso, pouco lazer"
                      },
                      "sono": {
                        "type": "string",
                        "example": "Dorme 5h por noite"
                      },
                      "atividadeFisica": {
                        "type": "string",
                        "example": "Raramente"
                      },
                      "estresse": {
                        "type": "string",
                        "example": "Alto"
                      },
                      "convivencia": {
                        "type": "string",
                        "example": "Boa"
                      },
                      "relacaoFamiliar": {
                        "type": "string",
                        "example": "Distante"
                      },
                      "apoioSocial": {
                        "type": "string",
                        "example": "Poucos amigos"
                      },
                      "nivelFelicidade": {
                        "type": "number",
                        "example": 4
                      },
                      "ansiedade": {
                        "type": "number",
                        "example": 8
                      },
                      "pensamentosNegativos": {
                        "type": "string",
                        "example": "Constantes"
                      },
                      "objetivoTerapia": {
                        "type": "string",
                        "example": "Melhorar crises de ansiedade"
                      },
                      "temasDelicados": {
                        "type": "string",
                        "example": "Relacionamentos"
                      },
                      "estiloAtendimento": {
                        "type": "string",
                        "example": "Acolhedor"
                      },
                      "observacoesFinais": {
                        "type": "string",
                        "example": "Paciente muito retraída"
                      },
                      "autorizacaoLGPD": {
                        "type": "boolean",
                        "example": true
                      },
                      "habilitado": {
                        "type": "boolean",
                        "example": false
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2025-01-10T12:34:56.000Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2025-01-10T14:21:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID do psicólogo não informado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "ID do psicólogo é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao buscar pré-pacientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": []
}